added sweetalert2 to delete list

+++ src/components/ButtonActionForFormsForMaryWithDescriptiveNamingConvention/ButtonActionForFormsForMaryWithDescriptiveNamingConvention.js
@@ -1,5 +1,7 @@
 import React, { Component } from 'react';
 import axios from 'axios'
+import Swal from 'sweetalert2'
+
 
 class ButtonActionForFormsForMaryWithDescriptiveNamingConvention extends Component{
 
@@ -7,16 +9,40 @@ class ButtonActionForFormsForMaryWithDescriptiveNamingConvention extends Compone
         axios.put(`/list/reset`)
             .then((response) => {
                 console.log(`reset list`);
-                this.props.getShoppingList();
+                this.props.getShoppingList();   
             })
     }
 
     deleteList = (event) => {
-        axios.delete(`/list/delete`)
-            .then((response) => {
-                console.log('deleted list');
-                this.props.getShoppingList();
-            })
+        Swal.fire({
+            title: 'Are you sure?',
+            text: 'You will not be able to recover this imaginary file!',
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonText: 'Yes, delete it!',
+            cancelButtonText: 'No, keep it'
+        }).then((result) => {
+            if (result.value) {
+                Swal.fire(
+                    'Deleted!',
+                    'Your secrets are safe with us.',
+                    'success'
+                )
+                axios.delete(`/list/delete`)
+                    .then((response) => {
+                        console.log('deleted list');
+                        this.props.getShoppingList();
+                    })
+                // For more information about handling dismissals please visit
+                // https://sweetalert2.github.io/#handling-dismissals
+            } else if (result.dismiss === Swal.DismissReason.cancel) {
+                Swal.fire(
+                    'Cancelled',
+                    'Shopping List Restored, thanks for Shopping SMart, shop S-Mart.',
+                    'error'
+                )
+            }
+        })
     }
 
     render(){
